+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New script to parse download requests from S3 log files for UMP books 
 - New script to update current UofM manifests with fulfill endpoints to replace pdf/epub urls
 - Updated README with appendix and additions to avaliable processes
+- New process to add fulfill endpoints to Limited Access manifests after the ClusterProcess has run
 ## Fixed
 
 ## 2024-03-21 -- v0.13.0